+++ SendBirdSDK.framework/Headers/SBDUserMessage.h
@@ -50,6 +50,16 @@
 @property (assign, nonatomic, readonly) SBDMessageRequestState requestState;
 
 /**
+ Represents target user ids to mention when success to send the message.
+ This value is valid only when the message is a pending message or failed message.
+ If the message is a succeeded message, see `mentionedUserIds`
+ 
+ @since 3.0.147
+ @see see `mentionedUserIds` when the message is a succeeded message.
+ */
+@property (strong, nonatomic, readonly, nonnull) NSArray<NSString *> *requestedMentionUserIds;
+
+/**
  Serializes message object.
  
  @return Serialized <span>data</span>.