+++ SendBirdSDK.framework/Headers/SendBirdSDK.h
@@ -7,7 +7,6 @@
 //
 
 #import <UIKit/UIKit.h>
-#import <CoreGraphics/CoreGraphics.h>
 
 // In this header, you should import all the public headers of your framework using statements like #import "PublicHeader"
 // v3.0